requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pytest>=8.0.0
pandas>=2.0.0
Pillow>=10.0.0
